Refactored protractor tests
