+++ mapstory/templates/_header.html
@@ -28,12 +28,12 @@
                             <img class="img-circle" src="{% avatar_url user 35 %}"/>
                         </a>
                     </li>
-                    <li class="dropdown"><a href="#" class="dropdown-toggle" data-toggle="dropdown">{{ user.username }}<b class="caret"></b></a>
+                    <li class="dropdown"><a href="#" id="id_userlink" class="dropdown-toggle" data-toggle="dropdown">{{ user.username }}<b class="caret"></b></a>
                         <ul class="dropdown-menu">
                             <li><a href="{% url 'profile_detail' slug=user.username %}">{% trans "Profile" %}</a></li>
                             <li class="divider"></li>
-                            <li><a href="{% url "edit_profile" user.username %}">{% trans "Edit Profile" %}</a></li>
-                            <li><a href="{% url "account_password" %}">{% trans "Change Password" %}</a> </li>
+                            <li><a href="{% url 'edit_profile' user.username %}">{% trans "Edit Profile" %}</a></li>
+                            <li><a href="{% url 'account_password' %}">{% trans "Change Password" %}</a> </li>
                             <li><a href="{% url 'profile_detail' slug=user.username %}?showNotifications=true">{% trans "Notification Settings" %}</a></li>
                             <li class="divider"></li>
                             <li><a href="{% url 'getpage' 'started' %}">{% trans "Help" %}</a></li>
