+++ mapstory/templates/_login_register.html
@@ -84,7 +84,7 @@
                                 <div class="col-md-3"></div>
                                 <div class="col-md-8 checkbox under-form-area">
                                     <p><input type="checkbox" value="" name="digest" checked> Receive the monthly digest of cool projects and community updates.</p>
-                                    <p><input type="checkbox" value="" ng-model="agreed">I agree with the <a href="http://wiki.mapstory.org/Terms_of_Service" target="_blank" class="new-tab">Terms and Conditions</a>.</p>
+                                    <p><input id="id_agreed" type="checkbox" value="" ng-model="agreed">I agree with the <a href="http://wiki.mapstory.org/Terms_of_Service" target="_blank" class="new-tab">Terms and Conditions</a>.</p>
                                 </div>
                             </div>
                             <div class="form-group">
