+++ mapstory/tests/e2e/pages/home.po.js
@@ -6,12 +6,12 @@
  */
 'use strict';
 
-require('./waitReady.js');
+require('../tools/waitReady.js');
 var path = require('path');
 // Upload paths need to be absolute or error.
-const testLayerFileRelative = '../sampledata/lewisandclarktrail.csv';
+const testLayerFileRelative = '../../sampledata/lewisandclarktrail.csv';
 const testLayerFile = path.resolve(__dirname, testLayerFileRelative);
-var wait_times = require('./wait_times');
+var wait_times = require('../tools/wait_times');
 var auth = require('./auth.po');
