+++ mapstory/tests/e2e/tools/screenHelper.js
