+++ mapstory/tests/e2e/tools/waitReady.js
